make client engine standalone singleton class instead of angular service. fix game loop.
to elaborate: i think i want to open source this as a game engine, and for that i would want to decouple the core features from Angular. Will eventually package the client up entirely outside of Angular and just make an example game using Angular for routing and modularity.

Refactored several properties and methods on client engine to instance properties that will be accessed from the shared instance

re: gameloop, the issue with the simulation was that I was passing in a calculated, variable value for fixedDeltaTime, when it should've been... fixed. now the simulation is silky smooth baby